move type logic to its own file
